requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pymupdf==1.23.3
pandas==2.1.1
openpyxl==3.1.2
python-multipart==0.0.6
pytesseract==0.3.10
python-bidi==0.4.2
Pillow==10.0.1
python-dotenv==1.0.0
typing-extensions==4.8.0
numpy==1.26.0 